+++ libs/ngx-editorjs-demo/src/lib/config/index.ts
@@ -1,10 +1,22 @@
-import { PluginConfig, PluginCode, PluginLink, PluginMarker, PluginSimpleImage } from '@tinynodes/ngx-editorjs-plugins';
+import { PluginConfig } from '@tinynodes/ngx-editorjs';
+import {
+  PluginCode,
+  PluginHeader,
+  PluginLink,
+  PluginList,
+  PluginMarker,
+  PluginParagraph,
+  PluginSimpleImage
+} from '@tinynodes/ngx-editorjs-plugins';
 
 /**
  * Creates the tools for the `ngx-tinynodes` demo application
  */
 export function createNgxEditorJSDemoTools(): PluginConfig {
   return {
+    parargraph: new PluginParagraph(),
+    header: new PluginHeader(),
+    list: new PluginList(),
     code: new PluginCode(),
     image: new PluginSimpleImage(),
     link: new PluginLink(),