+++ libs/ngx-editorjs/src/index.ts
@@ -4,3 +4,4 @@ export { NgxEditorJSModule } from './lib/ngx-editorjs.module';
 export { NgxEditorJSService } from './lib/services/editorjs.service';
 export { Block, HeaderBlock, ImageBlock, ListBlock, ListStyle, ParagraphBlock } from './lib/types/blocks';
 export { EditorJSConfig, NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './lib/types/config';
+export { BasePlugin, UserPlugins, PluginConfig } from './lib/types/plugins';